apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
gunicorn==22.0.0

pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.3.0

sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.14.0

python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0

structlog==24.4.0
orjson==3.10.7
